+++ kwap/src/logging.rs
@@ -1,6 +1,6 @@
 use core::fmt::Write;
 
-use kwap_common::prelude::*;
+use toad_common::prelude::*;
 use tinyvec::ArrayVec;
 
 use crate::platform;