+++ kwap/src/req/method.rs
@@ -1,4 +1,4 @@
-use kwap_msg::Code;
+use toad_msg::Code;
 
 use crate::code;